requirements.txt
streamlit==1.38.0
plotly==5.3.1
pandas==2.1.4
numerize==0.12
streamlit-option-menu==0.3.13
gspread==5.7.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
